api/requirements.txt
altgraph==0.17.2
aniso8601==8.0.0
apispec==5.1.1
autopep8==1.5.7
beautifulsoup4==4.10.0
captcha==0.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.1
colorama==0.4.4
docopt==0.6.2
Flask==1.1.2
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
future==0.18.2
geojson==2.5.0
gTTS==2.2.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Js2Py==0.71
json-excel-converter==1.3.0
MarkupSafe==1.1.1
marshmallow==3.14.1
MouseInfo==0.1.3
numpy==1.21.2
opencv-python==4.5.4.58
packaging==21.0
pefile==2021.9.3
pika==1.2.0
Pillow==8.3.2
pipwin==0.5.1
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycryptodome==3.11.0
pygame==2.0.1
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==0.10.1
pyowm==3.2.0
pyparsing==2.4.7
pyperclip==1.8.2
pypng==0.0.21
PyPrind==2.11.3
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
python-dotenv==0.15.0
python-vlc==3.0.12118
pytweening==1.0.4
pytz==2019.3
pywhatkit==5.1
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.23.0
rq==1.7.0
selenium==3.141.0
six==1.14.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tabulate==0.8.9
toml==0.10.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.25.11
waitress==2.0.0
webargs==8.0.1
Werkzeug==1.0.1
wikipedia==1.4.0
